Aliases, SeqMeta, SOL packager, qsub memory, bugfixes and more
added Aliases class to centrally deal with alternate column names

Files - retooled qsub to manage memeory better
	qsubMultiple / added numberOfJobs to force, a requirement for Itasca
	affects DeNovoCNV
	affects SkatMeta
	affects SkatMetaPrimary

SkatMeta is being updated and rebranded as SeqMeta,
	and our scripts were updated accordingly

Aliases - new class to centrally deal with alternate column names
	affects Hits
	affects SnpMarkerSet
	
SnpMarkerSet - fixed bug and adds support for 3col comma-delimited maps

HitWindows - adds support for the lookup of known loci

ChargeS - PLINK parser based on MAF

ResultsPackager - added support for the custom SOL program

Plink - batchLD calculations

PhenoPrep - ability to add additional columns, plus ID column

SkatMetaOutliers - test all possible combinations for SKEW+KURT

GenParser - better support for commas and for compressed files

HitWindows - first stab at a hits lookup

ext - fixed delimiter bug in getDate

+++ src/gwas/SkatMetaPrimary.java
@@ -121,14 +121,17 @@ public class SkatMetaPrimary {
 							+ "write.table(results, \"" + cohort + "_chr" + i + "_beforeSave_results.csv\", sep=\",\", row.names = F)\n"
 							+ "save(" + cohort + "_chr" + i + ", file=\"" + cohort + "_chr" + i + ".RData\", compress=\"bzip2\")";
 
-					filename = batchDir + cohort+ "_chr" + i + ".R";
-					out = new PrintWriter(new FileOutputStream(filename));
-					out.println(rCode);
-					out.close();
-
-					// then run the R code, if it only takes a few minutes then on your machine
-//					Runtime.getRuntime().exec("C:/Progra~1/R/R-3.0.1/bin/Rscript " + batchDir + "chr" + i + ".R");
-					v.add(filename);
+					
+					if (!Files.exists(resultDir+cohort + "_chr" + i + ".RData")) {
+						filename = batchDir + cohort+ "_chr" + i + ".R";
+						out = new PrintWriter(new FileOutputStream(filename));
+						out.println(rCode);
+						out.close();
+	
+						// then run the R code, if it only takes a few minutes then on your machine
+	//					Runtime.getRuntime().exec("C:/Progra~1/R/R-3.0.1/bin/Rscript " + batchDir + "chr" + i + ".R");
+						v.add(filename);
+					}
 					
 				}
 			}
@@ -140,6 +143,7 @@ public class SkatMetaPrimary {
 			}
 			
 			iterations = Matrix.toMatrix(Array.toStringArray(v));
+			System.out.println(iterations.length+"\tremaining to run for "+cohort);
 			if (System.getProperty("os.name").startsWith("Windows")) {
 				commands = "Rscript --no-save [%0]";
 				Files.batchIt(batchDir + "run", "", 5, commands, iterations);
@@ -148,6 +152,23 @@ public class SkatMetaPrimary {
 				commands = Rscript.getRscriptExecutable(new Logger())+" --no-save [%0]";
 //				Files.qsub("checkObject", dir, -1, commands, iterations, qsubMem, qsubWalltime);
 				Files.qsub(batchDir + "run_" + cohort, batchDir, -1, commands, iterations, qsubMem, qsubWalltime);
+				if (iterations.length == 0) {
+					new File(batchDir + "master.run_" + cohort).renameTo(new File(batchDir + "master.run_" + cohort+".bak"));
+				}
+			}
+			
+			if (iterations.length > 0) {
+				commands = "cd "+batchDir+"\n";
+				if (iterations.length == 1) {
+					commands += "./run_"+cohort+".qsub\n";
+				} else {
+					for (int j = 0; j < iterations.length; j++) {
+						commands += "./run_"+cohort+"_"+(j+1)+".qsub\n";
+					}
+				}
+				Files.qsub(batchDir + "finishUpOnSB_" + cohort, commands, 60000, (int)Math.ceil((double)iterations.length/2.0), 16);
+			} else {
+				new File(batchDir + "finishUpOnSB_" + cohort).delete();
 			}
 
 			v = new Vector<String>();
@@ -159,7 +180,7 @@ public class SkatMetaPrimary {
 				v.add("load(\"" + ext.rootOf(iterations[i][0])+ ".RData\")");
 				consolidate += (i==0?"":", ")+ext.rootOf(iterations[i][0]);
 				jobNamesWithAbsolutePaths.add(batchDir + "run_" + cohort+"_"+(i+1)+".qsub");
-				jobSizes.add((int)new File(ext.replaceAllWith(genos, "#", (i+1)+"")).length());
+				jobSizes.add(Integer.MAX_VALUE - (int)new File(ext.replaceAllWith(genos, "#", (i+1)+"")).length());
 			}
 			consolidate += ")";
 			v.add(consolidate);
@@ -167,8 +188,9 @@ public class SkatMetaPrimary {
 			v.add("save(" + cohort + ", file=\"" + cohort + ".RData\", compress=\"bzip2\")");
 			Files.writeList(Array.toStringArray(v), batchDir + "mergeRdataFiles.R");
 			commands = Rscript.getRscriptExecutable(new Logger())+" --no-save "+batchDir + "mergeRdataFiles.R";
-			Files.qsub(batchDir + "run_mergeRdataFiles_" + cohort, commands, qsubMem*4, qsubWalltime);
-			Files.qsubMultiple(jobNamesWithAbsolutePaths, jobSizes, batchDir, "chunks", 8, true, 2000, 1);
+			Files.qsub(batchDir + "run_mergeRdataFiles_" + cohort, commands, qsubMem*4, qsubWalltime, 1);
+			Files.qsubMultiple(jobNamesWithAbsolutePaths, jobSizes, batchDir, batchDir+"chunk_"+cohort, 8, true, null, -1, qsubMem, qsubWalltime);
+			Files.qsubMultiple(jobNamesWithAbsolutePaths, jobSizes, batchDir, batchDir+"chunkSB_"+cohort, 16, true, "sb256", -1, qsubMem, qsubWalltime);
 
 		} catch (IOException e) {
 			e.printStackTrace();
@@ -197,7 +219,34 @@ public class SkatMetaPrimary {
 			}
 		}
 		Files.writeList(Array.toStringArray(v), "scriptAll");
+		Files.chmod("scriptAll");
 
+		v = new Vector<String>();
+		for (int i = 0; i < phenos.length; i++) {
+			for (int j = 0; j < races.length; j++) {
+				v.add("cd "+cohort+"_"+races[j]+"_"+phenos[i]+"/batchFiles/");
+				v.add("./master.chunk_"+cohort+"_"+races[j]+"_"+phenos[i]);
+				v.add("cd ../../");
+				v.add("");
+			}
+		}
+		Files.writeList(Array.toStringArray(v), "scriptAllItasca");
+		Files.chmod("scriptAllItasca");
+
+		v = new Vector<String>();
+		for (int i = 0; i < phenos.length; i++) {
+			for (int j = 0; j < races.length; j++) {
+				if (Files.exists(cohort+"_"+races[j]+"_"+phenos[i]+"/batchFiles/finishUpOnSB_" +cohort+"_"+races[j]+"_"+phenos[i])) {
+					v.add("cd "+cohort+"_"+races[j]+"_"+phenos[i]+"/batchFiles/");
+					v.add("qsub -q sb finishUpOnSB_" +cohort+"_"+races[j]+"_"+phenos[i]);
+					v.add("cd ../../");
+					v.add("");
+				}
+			}
+		}
+		Files.writeList(Array.toStringArray(v), "finishUpOnSB");
+		Files.chmod("finishUpOnSB");
+		
 		v = new Vector<String>();
 		for (int i = 0; i < phenos.length; i++) {
 			for (int j = 0; j < races.length; j++) {
@@ -209,6 +258,185 @@ public class SkatMetaPrimary {
 			}
 		}
 		Files.writeList(Array.toStringArray(v), "mergeAll");
+		Files.chmod("mergeAll");
+
+		v = new Vector<String>();
+		for (int i = 0; i < phenos.length; i++) {
+			for (int j = 0; j < races.length; j++) {
+				v.add("cd "+cohort+"_"+races[j]+"_"+phenos[i]+"/results/");
+				v.add("tar -zcvf ../../"+cohort+"_"+races[j]+"_"+phenos[i]+"_"+ext.getDate(new Date(), "")+".tar.gz "+cohort+"_"+races[j]+"_"+phenos[i]+"*.RData");
+				v.add("cd ../../");
+				v.add("");
+			}
+		}
+		Files.writeList(Array.toStringArray(v), "packageUpAll");
+		Files.chmod("packageUpAll");
+	}
+
+	public static void additionalModels(String cohort, String phenosCommaDelimited, String snpInfo, int qsubMem, double qsubWalltime) {
+		String[] phenos;
+		Vector<String> v;
+		
+		phenos = phenosCommaDelimited.split(",");
+		
+		v = new Vector<String>();
+		for (int i = 0; i < phenos.length; i++) {
+			try {
+				batchAdditionals(phenos[i], cohort, snpInfo, qsubMem, qsubWalltime);
+			} catch (Exception e) {
+				System.err.println("Error - failed to script up "+phenos[i]);
+			}
+			v.add("cd "+phenos[i]+"/batchFiles/");
+			v.add("./master.run_additionals");
+			v.add("cd ../../");
+			v.add("");
+		}
+		Files.writeList(Array.toStringArray(v), "addAll");
+		Files.chmod("addAll");
+	}
+
+	private static void batchAdditionals(String phenoDirectory, String cohort, String snpInfo, int qsubMem, double qsubWalltime) {
+		String phenoDir;
+		String phenoRoot;
+		String resultDir;
+		String currentGeno;
+		String currentSnpInfo;
+		String rCode;
+		String batchDir;
+		PrintWriter out;
+		Vector<String> v;
+		String filename, commands;
+		String[][] iterations;
+		boolean foundGenos;
+		boolean foundSnpInfo;
+		String consolidate;
+		Vector<String> jobNamesWithAbsolutePaths;
+		IntVector jobSizes;
+
+		if (! new File(phenoDirectory).exists()) {
+			System.err.println("Error - directory not found: " + phenoDirectory);
+			return;
+		}
+
+		// create directory called dir+root+"/"; example: "c:/diffpath/pheno_F7_studyIDs/"
+		phenoDir = ext.verifyDirFormat(ext.parseDirectoryOfFile(new File(phenoDirectory).getAbsolutePath()));
+		resultDir = phenoDir + "/results/";
+		batchDir = phenoDir + "/batchFiles/";
+
+		try {
+			v = new Vector<String>();
+			// generate batch files in dir+root+"/batchFiles/"; example: "c:/diffpath/pheno_F7_studyIDs/batchFiles/chr1.R"
+			foundGenos = false;
+			foundSnpInfo = false;
+			for (int i = 1; i <= 26; i++) {
+				if (snpInfo.contains("#")) {
+					currentSnpInfo = ext.replaceAllWith(snpInfo, "#", i+"");
+				} else {
+					currentSnpInfo = snpInfo;
+				}
+				if (new File(resultDir+cohort+"_chr"+i+".RData").exists()) {
+					foundGenos = true;
+					foundSnpInfo = true;
+					rCode = "library(\"skatMeta\")\n"
+							+ "setwd(\"" + resultDir + "\")\n"
+							+ "\n"
+							+(currentSnpInfo.toLowerCase().endsWith(".rdata")
+									?"obj_name <- load(\"" + currentSnpInfo + "\")\n"
+									+"SNPInfo <- get(obj_name)\n"
+											+"rm(list=obj_name)\n"
+											+"rm(obj_name)\n"
+									:"SNPInfo <- read.csv(\"" + currentSnpInfo + "\", header=T, as.is=T)\n"
+									)
+							+ "\n"
+							
+							+"cohortName <- load(\"" + resultDir+cohort+"_chr"+i+".RData" + "\")\n"
+							+"cohort <- get(cohortName)\n"
+
+//							+ "results <- singlesnpMeta(cohort, SNPInfo=SNPInfo, snpNames = \"Name\", cohortBetas = TRUE)\n"
+//							+ "write.table(results, \"" + cohort + "_chr" + i + "_SingleSnp_results.csv\", sep=\",\", row.names = F)\n"
+//							+ "results <- burdenMeta(" + cohort + "_chr" + i + ", aggregateBy=\"SKATgene\", mafRange = c(0,0.05), SNPInfo=subset(SNPInfo, sc_nonsynSplice==TRUE), snpNames=\"SNP\", wts = 1)\n"
+//							+ "write.table(results, \"" + cohort + "_chr" + i + "burden_results.csv\", sep=\",\", row.names = F)\n"
+//							+ "results <- singlesnpMeta(cohort, SNPInfo=SNPInfo, snpNames = \"Name\", cohortBetas = TRUE)\n"
+//							+ "write.table(results, \"" + cohort + "_chr" + i + "_SingleSnp_results.csv\", sep=\",\", row.names = F)\n"
+
+							+ "results <- singlesnpMeta(cohort, SNPInfo=SNPInfo_ExomeChipV5, snpNames = \"Name\", aggregateBy=\"SKATgene\", cohortBetas = TRUE)\n"
+							+ "write.table( results, \"SingleSNP_chr\"+(i)+\".csv\", sep=\",\", row.names = F)\n"
+
+							+ "results <- burdenMeta(cohort, SNPInfo=subset(SNPInfo_ExomeChipV5, sc_nonsynSplice==TRUE), snpNames = \"Name\", aggregateBy=\"SKATgene\", mafRange = c(0,0.01), wts = 1)\n"
+							+ "write.table( results, \"T1Count_chr\"+(i)+\".csv\", sep=\",\", row.names = F)\n"
+
+							+ "results <- burdenMeta(cohort, SNPInfo=subset(SNPInfo_ExomeChipV5, sc_nonsynSplice==TRUE), snpNames = \"Name\", aggregateBy=\"SKATgene\", mafRange = c(0,0.05), wts = 1)\n"
+							+ "write.table( results, \"T5Count_chr\"+(i)+\".csv\", sep=\",\", row.names = F)\n"
+
+							+ "results <- skatMeta(cohort, SNPInfo=subset(SNPInfo_ExomeChipV5, sc_nonsynSplice==TRUE), snpNames = \"Name\", aggregateBy=\"SKATgene\", mafRange = c(0,0.01), wts = 1)\n"
+							+ "write.table( results, \"SKAT_T1_chr\"+(i)+\".csv\", sep=\",\", row.names = F)\n"
+
+							+ "results <- skatMeta(cohort, SNPInfo=subset(SNPInfo_ExomeChipV5, sc_nonsynSplice==TRUE), snpNames = \"Name\", aggregateBy=\"SKATgene\", mafRange = c(0,0.05), wts = 1)\n"
+							+ "write.table( results, \"SKAT_T5_chr\"+(i)+\".csv\", sep=\",\", row.names = F)\n"
+
+							
+							;
+
+					
+					filename = batchDir + "additionals_chr" + i + ".R";
+					out = new PrintWriter(new FileOutputStream(filename));
+					out.println(rCode);
+					out.close();
+
+					v.add(filename);
+					
+				}
+			}
+
+			iterations = Matrix.toMatrix(Array.toStringArray(v));
+			System.out.println(iterations.length+"\tremaining to run for "+cohort);
+			if (System.getProperty("os.name").startsWith("Windows")) {
+				commands = "Rscript --no-save [%0]";
+				Files.batchIt(batchDir + "run", "", 5, commands, iterations);
+			} else {
+				commands = Rscript.getRscriptExecutable(new Logger())+" --no-save [%0]";
+				Files.qsub(batchDir + "run_additionals", batchDir, -1, commands, iterations, qsubMem, qsubWalltime);
+				if (iterations.length == 0) {
+					new File(batchDir + "master.run_additionals").renameTo(new File(batchDir + "master.run_additionals.bak"));
+				}
+			}
+
+//			v = new Vector<String>();
+//			jobNamesWithAbsolutePaths = new Vector<String>();
+//			jobSizes = new IntVector();
+//			v.add("setwd(\"" + resultDir + "\")");
+//			consolidate = cohort + "<- c(";
+//			for (int i = 0; i < iterations.length; i++) {
+//				v.add("load(\"" + ext.rootOf(iterations[i][0])+ ".RData\")");
+//				consolidate += (i==0?"":", ")+ext.rootOf(iterations[i][0]);
+//				jobNamesWithAbsolutePaths.add(batchDir + "run_" + cohort+"_"+(i+1)+".qsub");
+//				jobSizes.add((int)new File(ext.replaceAllWith(genos, "#", (i+1)+"")).length());
+//			}
+//			consolidate += ")";
+//			v.add(consolidate);
+//			v.add("class("+cohort + ") <- \"skatCohort\"");
+//			v.add("save(" + cohort + ", file=\"" + cohort + ".RData\", compress=\"bzip2\")");
+//			Files.writeList(Array.toStringArray(v), batchDir + "mergeRdataFiles.R");
+//			commands = Rscript.getRscriptExecutable(new Logger())+" --no-save "+batchDir + "mergeRdataFiles.R";
+//			Files.qsub(batchDir + "run_mergeRdataFiles_" + cohort, commands, qsubMem*4, qsubWalltime);
+//			Files.qsubMultiple(jobNamesWithAbsolutePaths, jobSizes, batchDir, batchDir+"chunk_"+cohort, 8, true, 2000, 1);
+
+		} catch (IOException e) {
+			e.printStackTrace();
+		}
+	}
+
+	public static void renameMany(String cohort, String phenosCommaDelimited, String racesCommaDelimited, String insert) {
+		String[] phenos, races;
+		
+		phenos = phenosCommaDelimited.split(",");
+		races = racesCommaDelimited.split(",");
+		
+		for (int i = 0; i < phenos.length; i++) {
+			for (int j = 0; j < races.length; j++) {
+				new File(cohort+"_"+races[j]+"_"+phenos[i]+".RData").renameTo(new File(cohort+insert+"_"+races[j]+"_"+phenos[i]+".RData"));
+			}
+		}
 	}
 
 	public static void main(String[] args) {
@@ -220,6 +448,8 @@ public class SkatMetaPrimary {
 		String races = "EA,AA";
 		int qsubMem;
 		double qsubWalltime;
+		boolean additionals = false;
+		String rename = null;
 
 		cohort="aric";
 		genos="D:/SkatMeta/genotypes_blacks_AA/AA_ARIC_noJHS_chr#t.csv";
@@ -240,6 +470,9 @@ public class SkatMetaPrimary {
 				"   (8) (optional) races to run (i.e. races="+races+" (default))\n"+
 				" ( the final optional parameters require a fixed phenotype format of [cohort name]_[race]_[phenotype]_.csv )\n"+
 				" ( as well as a genotype file with the format of regular_filename_[%race]_chr#.csv )\n"+
+				"   (9) (optional) run additional models for Rdata files (i.e. -additionals  (not the default))\n"+
+				" OR\n"+
+				"   (9) (optional) rename Rdata files with this insert (i.e. rename=42PCs (not the default))\n"+
 				"";
 
 		for (int i = 0; i<args.length; i++) {
@@ -262,13 +495,23 @@ public class SkatMetaPrimary {
 				phenos = args[i].split("=")[1];
 			} else if (args[i].startsWith("races=")) {
 				races = args[i].split("=")[1];
+			} else if (args[i].startsWith("-additionals")) {
+				additionals = true;
+			} else if (args[i].startsWith("rename=")) {
+				rename = args[i].split("=")[1];
 			} else {
 				System.err.println("Error - invalid argument: "+args[i]);
 			}
 		}
 		
 		if (phenos != null) {
-			batchMany(cohort, genos, phenos, races, snpInfo, qsubMem, qsubWalltime);
+			if (rename != null) {
+				renameMany(cohort, phenos, races, rename);
+			} else if (additionals) {
+				additionalModels(cohort, phenos, snpInfo, qsubMem, qsubWalltime);
+			} else {
+				batchMany(cohort, genos, phenos, races, snpInfo, qsubMem, qsubWalltime);
+			}
 		} else {
 			batch(cohort, genos, pheno, snpInfo, qsubMem, qsubWalltime);
 		}
